refactor(organon): extract energeia::shared submodule (#2642)
Move EnergeiaServices struct and helper functions (require_str, opt_str,
opt_u64, opt_bool, to_json_text) to shared.rs. Re-export EnergeiaServices
from mod.rs to preserve the public path.
